python-backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "python-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "kevin.krausse",email = "kevin.krausse@taxbit.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
